Adds `lf hid pack` and `lf hid unpack` offline commands.
These are used to pack and unpack the bit length, facility code and card number from "short" HID Prox IDs.

This also simplifies some repeated code that is used to parse hex digits in lfhid, and adds basic client-side validation to the `lf hid clone` and `lf hid sim` commands, for when an ID that is too long is entered.