crypto1_create: initialize malloc()ed memory to 0
On systems that support malloc(), crypto1_create() allocates a new
struct Crypto1State and starts calculations without initializing the
allocated memory area.

At least on Linux, malloc(3) says that the allocated memory is not
initialized. Looking at the version for bare metal ARM, it seems that
the components of struct Crypto1State should be 0 initially. Add a
memset() call to initialize them.